backend/requirements.txt
# PhishGuard Backend Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1

# Machine Learning & NLP
scikit-learn==1.3.2
spacy==3.7.2
nltk==3.8.1
numpy==1.25.2
pandas==2.1.4

# Email Processing
email-validator==2.1.0
python-email-validator==2.0.0
beautifulsoup4==4.12.2
lxml==4.9.3

# HTTP & External APIs
httpx==0.25.2
requests==2.31.0
aiofiles==23.2.1

# Background Tasks & Async
celery==5.3.4
redis==5.0.1

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0

# Development
python-dotenv==1.0.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# WebSocket Support
websockets==12.0
python-socketio==5.10.0

# File Processing
python-magic==0.4.27
openpyxl==3.1.2

# Date/Time
python-dateutil==2.8.2

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS
fastapi-cors==0.0.6 